Update Mystcraft API

+++ resources/com/xcompwiz/mystcraft/api/internals/IFormula.java
@@ -1,8 +1,20 @@
 
 package com.xcompwiz.mystcraft.api.internals;
 
+/**
+ * A basic algebra math formula interface
+ * 
+ * @author xcompwiz
+ */
 public interface IFormula {
 
-    public abstract float calc(int x);
+    /**
+     * Returns a value based on the value of x
+     * 
+     * @param x
+     *        The x position
+     * @return The y position
+     */
+    public abstract float calc(float x);
 
 }
